login v2

+++ models/models.go
@@ -2,7 +2,7 @@ package models
 
 // Users struct
 type Users struct {
-	ID       int    `json:"id" db:"id"`
+	ID       int64  `json:"id" db:"id"`
 	Username string `json:"username" db:"username"`
 	Password string `json:"password" db:"password"`
 	Email    string `json:"email" db:"email"`
